% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioreg.R
\docType{data}
\name{bioreg}
\alias{bioreg}
\title{Biorégions.}
\format{
Fichier géoréférencé au format sf contenant 3 biorégions.
}
\source{

}
\usage{
bioreg
}
\arguments{
\item{LIBELLE}{= nom de la région}

\item{geometry}{= géométrie}
}
\description{
Carte des biorégions
}
\examples{
data(bioreg)
plot(st_geometry(bioreg), border='blue', lwd=2)
plot(st_geometry(greco), add=T, border='red')

}
\keyword{datasets}
