% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodesEssIFN.R
\docType{data}
\name{CodesEssIFN}
\alias{CodesEssIFN}
\title{Codification des essences par l'IFN.}
\format{
data frame contenant près de 142 lignes et 2 variables
}
\source{
\url{https://inventaire-forestier.ign.fr}
}
\usage{
CodesEssIFN
}
\arguments{
\item{code}{= code de l'IFN}

\item{libelle}{= Nom vernaculaire}
}
\description{
Table permettant de faire la correespondance entre le code IFN et le nom vernaculaire
}
\examples{
data(CodesEssIFN)

}
\keyword{datasets}
