% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodesIFNmod.R
\docType{data}
\name{CodesIFNmod}
\alias{CodesIFNmod}
\title{Codes IFN.}
\format{
Data frame contenant 117 lignes et 3 variables.
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?article729}
}
\usage{
CodesIFNmod
}
\arguments{
\item{Code}{= nom de la variable IFN contenue dans la table IFNarbres ou IFNplacettes}

\item{Modalité}{= modalité}

\item{Libelle}{= definition de la modalité}
}
\description{
Modalités des variables arbres et placettes de l'IFN
}
\examples{
data(CodesIFNmod)
 
}
\keyword{datasets}
