% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoeffInflation.R
\name{CoeffInflation}
\alias{CoeffInflation}
\title{Tableau des coefficients d'inflation}
\format{
Data frame contenant 70 lignes et 3 variables.
}
\source{
\url{https://www.insee.fr/fr/statistiques/serie/010605954#Tableau}
}
\usage{
CoeffInflation()
}
\description{
Fonction permettant de créer un tableau des coefficients d'inflation.
}
\examples{
CoeffInflation()
 
}
