% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeadEssGreco.R
\name{DeadEssGreco}
\alias{DeadEssGreco}
\title{Mortalité par essence et GRECO}
\source{
IFN
}
\usage{
DeadEssGreco(ess)
}
\arguments{
\item{ess}{= code essence de l'IFN. Le code data("CodesEssIFN") rappelle la codification de l'IFN.}
}
\value{
La fonction renvoie le taux de mortalité sous forme de tableaux ou de cartes.
}
\description{
La fonction DeadEssGreco fournit par grande région écologique (GRECO) le taux de mortalité par essence.
Il est définit comme le ratio en volume des arbres morts sur pied par rapport aux arbres vivants.
}
\examples{
ess = "05"
res = DeadEssGreco(ess)
# Part de l'essence dans la mortalité
res$gpart
# Taux de mortalité
res$gmort
# Maximum de mortalité
res$tab \%>\% 
 group_by(greco) \%>\% 
 arrange(desc(PourcM)) \%>\% 
 slice(1)
 # carte de la part de l'essence dans la mortalité
 res$mappart
 # carte du taux de mortalité de l'essence
 res$mapmort

}
\author{
Max Bruciamacchie
}
