% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeadEssSer.R
\name{DeadEssSer}
\alias{DeadEssSer}
\title{Mortalité par essence et sylvoécorégions}
\format{
La fonction renvoie le taux de mortalité sous forme de tableau ou de cartes.
}
\source{
IFN
}
\usage{
DeadEssSer(ess)
}
\arguments{
\item{ess}{= code essence de l'IFN. Le code data("CodesEssIFN") rappelle la codification de l'IFN.}
}
\description{
La fonction DeadEssSer fournit par sylvoécorégions le taux de mortalité par essence. 
Il est défini comme le le ratio en volume des arbres morts sur pied par rapport aux arbres vivants, 
sur les deux dernières périodes de 5 ans.
}
\examples{
ess = "05"
res = DeadEssSer(ess)
# sous forme tableau
head(res$tab)
# cartes du taux de mortablité sur les 2 dernières années
res$mappart

}
\author{
Max Bruciamacchie
}
