% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decChaude.R
\docType{data}
\name{decChaude}
\alias{decChaude}
\title{Décroissances métriques des tarifs Chaudé.}
\format{
Table contenant 19 lignes et 21 variables

\describe{
  \item{Classe}{Classe de diamètre}
  \item{Ch1 à Ch20}{20 numéros de tarifs Chaudé}
}
}
\source{
Vade-mecum du forestier
}
\usage{
decChaude
}
\description{
Coefficients des décroissances métriques variables pour les tarifs de cubage Chaudé à 2 entrées
}
\examples{
data(decChaude)

}
\keyword{datasets}
