% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecorces.R
\docType{data}
\name{Ecorces}
\alias{Ecorces}
\title{Epaisseur d'écorce - source IFN.}
\format{
Fichier contenant près de 64 lignes et 4 variables
}
\source{
\url{https://inventaire-forestier.ign.fr}
}
\usage{
Ecorces
}
\arguments{
\item{code}{= code essence de l'IFN}

\item{libelle}{= nom vernaculaire}

\item{a}{= ordonnée à l'origine}

\item{b}{= pente}
}
\description{
Coefficients de la relation linéaire entre le diamètre de l'arbre 
et son épaisseur d'écorce
}
\examples{
data(Ecorces)

}
\keyword{datasets}
