% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD.R
\docType{data}
\name{FD}
\alias{FD}
\title{Périmètres}
\format{
Fichier géoréférencé au format sf avec la projection Lambert93 (EPSG=2154).

\describe{
  \item{IIDTN_FRT}{identifiant de la forêt}
  \item{LLIB_FRT}{Nom de la forêt}
  \item{geometry}{géométrie}
}
}
\source{
\url{https://www.data.gouv.fr/fr/datasets/onf-contour-des-forets-publiques/}
}
\usage{
FD
}
\description{
Limites de 4 forêts domaniales
}
\examples{
data(FD)

}
\keyword{datasets}
