% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greco.R
\docType{data}
\name{greco}
\alias{greco}
\title{Grandes Régions Ecologiques (GRECO).}
\format{
Fichier géoréférencé au format sf contenant les 11 grandes régions écologiques
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?article729}
}
\usage{
greco
}
\arguments{
\item{greco}{= identifiant de la GRECO}

\item{geometry}{= géométrie}
}
\description{
Carte des Grandes Régions Ecologiques
}
\examples{
data(ser)
plot(st_geometry(greco), border='blue', lwd=2)
plot(st_geometry(ser), add=T, border='red')
 
}
\keyword{datasets}
