% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNarbres.R
\docType{data}
\name{IFNarbres}
\alias{IFNarbres}
\title{Fichier arbres de l'IFN.}
\format{
Data frame contenant près de 1 million de lignes et 14 variables
}
\source{
\url{https://inventaire-forestier.ign.fr}
}
\usage{
IFNarbres
}
\arguments{
\item{idp}{= identifiant du point d’inventaire}

\item{a}{= identifiant de l’arbre}

\item{espar}{= espèce arborée}

\item{veget}{= origine de l’arbre}

\item{mortb}{= mortalité de branches dans le houppier}

\item{acci}{= accident de l'arbre}

\item{ori}{= origine de l’arbre}

\item{c13}{= circonférence à 1,30 m (cm)}

\item{ir5}{= accroissement radial sous écorce sur 5 ans (mm)}

\item{htot}{= hauteur totale (m)}

\item{hdec}{= hauteur à la découpe (m)}

\item{v}{= volume de l'arbre (donnée calculée)}

\item{w}{= coefficient de pondération de l’arbre (donnée calculée)}

\item{Annee}{= année de mesure}
}
\description{
Base Arbres de l'IFN contenant près d'un million de lignes et 14 variables.
 Les modalités des variables espar, veget, mortb, acci et ori se trouvent 
 dans la table CodesIFNmod. Les modalités de la variable espar se trouvent dans la table
 CodesEssIFN.
}
\examples{
data(IFNarbres)

}
\keyword{datasets}
