% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNarbres_morts.R
\docType{data}
\name{IFNarbres_morts}
\alias{IFNarbres_morts}
\title{Fichier arbres de l'IFN.}
\format{
Data frame contenant pas moins de 50 000 lignes et 10 variables
}
\source{
\url{https://inventaire-forestier.ign.fr}
}
\usage{
IFNarbres_morts
}
\arguments{
\item{idp}{= identifiant du point d’inventaire}

\item{a}{= identifiant de l’arbre}

\item{espar}{= espèce arborée}

\item{veget}{= origine de l’arbre}

\item{datemort}{= date présumée de la mort}

\item{ori}{= origine de l’arbre}

\item{c13}{= circonférence à 1,30 m (cm)}

\item{v}{= volume de l'arbre (donnée calculée)}

\item{w}{= coefficient de pondération de l’arbre (donnée calculée)}

\item{Annee}{= année de mesure}
}
\description{
Base arbres morts de l'IFN inventoriés sur les placettes forestières contenant pas moins de 50 000 lignes et 10 variables.
 Les modalités des variables espar, veget, datemort et ori se trouvent 
 dans la table CodesIFNmod. Les modalités de la variable espar se trouvent dans la table
 CodesEssIFN.
}
\examples{
data(IFNarbres_morts)

}
\keyword{datasets}
