% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNdata.R
\encoding{UTF-8}
\name{IFNdata}
\alias{IFNdata}
\title{Rechargement des donnees IFN}
\usage{
IFNdata(enrg = TRUE)
}
\arguments{
\item{enrg}{= si TRUE les 2 tables Arbres et Placettes sont enregistrées sous forme d'archives rda.
Sinon la fonction renvoie les deux tables.}
}
\value{
Cette fonction enregistre dans le dossier data les deux tables IFNarbres.rda et IFNplacettes.rda.
}
\description{
Cette fonction permet de télécharger les données brutes de l'IFN à partir du site
https://inventaire-forestier.ign.fr/IMG/zip/. Elle charge les données de 2005 à 2018.
}
\examples{
IFNdata()
# ou bien
res <- IFNdata(FALSE)
IFNarbres <- res$IFNarbres
IFNplacettes <- res$IFNplacettes

}
\author{
BRUCIAMACCHIE Max
}
