% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSEEcom.R
\docType{data}
\name{INSEEcom}
\alias{INSEEcom}
\title{Communes françaises.}
\format{
Data frame contenant 35416 lignes et 12 variables.
}
\source{
\url{https://www.insee.fr/fr/information/2666684#titre-bloc-26}
}
\usage{
INSEEcom
}
\arguments{
\item{CDC}{= Découpage de la commune en cantons}

\item{CHEFLIEU}{= Chef-lieu d'arrondissement, de département, de région ou bureau centralisateur de canton}

\item{REG}{= Code région}

\item{DEP}{= Code département}

\item{COM}{= Code commune}

\item{AR}{= Code arrondissement}

\item{CT}{= Code canton}

\item{TNCC}{= Type de nom en clair}

\item{ARTMAJ}{= Article (majuscules)}

\item{NCC}{= Nom en clair (majuscules)}

\item{ARTMIN}{= Article (typographie riche)}

\item{NCCENR}{= Nom en clair (typographie riche)}
}
\description{
Liste des communes françaises existantes au 1er janvier 2017.
Cette table permet la correspondance avec les régions, départements, cantons et arrondissements.
}
\examples{
data(INSEEcom)
 
}
\keyword{datasets}
