% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSEEdep.R
\docType{data}
\name{INSEEdep}
\alias{INSEEdep}
\title{Départements français.}
\format{
Data frame contenant 101 lignes et 5 variables.
}
\source{
\url{https://www.insee.fr/fr/information/2666684#titre-bloc-26}
}
\usage{
INSEEdep
}
\arguments{
\item{REGION}{= Code région}

\item{DEP}{= Code département}

\item{CHEFLIEU}{= Chef-lieu d'arrondissement, de département, de région ou bureau centralisateur de canton}

\item{TNCC}{= Type de nom en clair}

\item{NCC}{= Nom en clair (majuscules)}

\item{NCCENR}{= Nom en clair (typographie riche)}
}
\description{
Liste des départemets français existants au 1er janvier 2017.
Cette table permet la correspondance avec les régions.
}
\examples{
data(INSEEdep)
 
}
\keyword{datasets}
