% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSEEpa.R
\docType{data}
\name{INSEEpa}
\alias{INSEEpa}
\title{Pouvoir d'achat.}
\format{
Data frame contenant 70 lignes et 3 variables.
}
\source{
\url{https://www.insee.fr/fr/information/2417794}
}
\usage{
INSEEpa
}
\arguments{
\item{Année}{= année. La base démarre à l'année 1950}

\item{Infla}{= inflation. L'inflation annuelle se calcule comme le ratio de 2 années}

\item{Monnaie}{= variable prennant en compte le passage des anciens francs aux nouveaux francs en 1960 
et le passage des francs aux euros en 2002 (1 € = 6,55957 FF).}
}
\description{
Pouvoir d'achat d'une somme en euros ou en francs d’une année donnée en une somme 
équivalente en euros ou en francs d’une autre année, corrigée de l’inflation observée
 entre les deux années.
}
\examples{
data(INSEEpa)
 
}
\keyword{datasets}
