% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParFor.R
\docType{data}
\name{ParFor}
\alias{ParFor}
\title{Exemple de parcellaire forestier}
\format{
Fichier géoréférencé au format sf avec la projection Lambert93 (EPSG=2154).
}
\source{
\url{https://www.data.gouv.fr/fr/datasets/onf-contour-des-forets-publiques/}
}
\usage{
ParFor
}
\arguments{
\item{NumPar}{= numéro de la parcelle}

\item{geometry}{= géométrie}
}
\description{
Cas de la FD de Montargis
}
\examples{
data(ParFor)

}
\keyword{datasets}
