% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanComptable.R
\docType{data}
\name{PlanComptable}
\alias{PlanComptable}
\title{Plan comptable.}
\format{
Data frame contenant 40 lignes et 3 variables.
}
\source{
Issu de l'AFI (association futaie irrégulière)
}
\usage{
PlanComptable
}
\arguments{
\item{Compte}{= Codification respectant le plan comptable national.}

\item{Intitulé}{= Intitulé en clair}

\item{Détail}{= détails facilitant l'attribution du code}
}
\description{
Plan comptable permettant un suivi des recettes et dépenses.
}
\examples{
data(PlanComptable)
 
}
\keyword{datasets}
