% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtectDataCreate.R
\name{ProtectDataCreate}
\alias{ProtectDataCreate}
\title{Base de données statuts de protection}
\source{
https://inpn.mnhn.fr/docs/Shape
}
\usage{
ProtectDataCreate()
}
\value{
La fonction crée un objet géoréférencé au format sf.
}
\description{
La fonction crée un objet qui rassemble tous les statuts de protections disponibles sur
le serveur du Muséum d'Histoire naturelle.
}
\examples{
library(tidyverse)
library(readxl)
library(tools)
library(sf)

Protect <- ProtectDataCreate()

}
\author{
Bruciamacchie Max
}
