% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnIFN.R
\docType{data}
\name{rnIFN}
\alias{rnIFN}
\title{Régions naturelles de l'IFN.}
\format{
Fichier géoréférencé contenant 309 régions naturelles

\describe{
  \item{REGN}{identifiant de la région naturelle}
  \item{REGIONN}{intitulé de la région naturelle}
  \item{geometry}{géométrie}
}
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?article729}
}
\usage{
rnIFN
}
\description{
Régions naturelles de l'IFN.
}
\examples{
data(rnIFN)
 
}
\keyword{datasets}
