% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser.R
\docType{data}
\name{ser}
\alias{ser}
\title{Sylvoécorégions.}
\format{
Fichier géoréférencé contenant 86 sylvoécorégions.

\describe{
  \item{codeser}{code de la sylvoécorégion}
  \item{NomSER}{Nom de la sylvoécorégion}
  \item{greco}{Nom de la grande région écologique}
  \item{geometry}{géométrie}
}
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?article729}
}
\usage{
ser
}
\description{
Carte des sylvoécorégions
}
\examples{
data(ser)
plot(st_geometry(ser))

}
\keyword{datasets}
