% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tarif_EMERGE.R
\name{Tarif_EMERGE}
\alias{Tarif_EMERGE}
\title{Volume total d'un arbre selon le tarif de cubage du projet EMERGE}
\format{
Renvoie le data frame avec une nouvelle colonne VTot
}
\source{
ONF
}
\usage{
Tarif_EMERGE(df, Haut, C130)
}
\arguments{
\item{df}{= data frame contenant des informations sur les essences, hauteurs et circonférences}

\item{Haut}{= variable de la hauteur totale en m}

\item{C130}{= circonférence mesurée à 1,30m en cm}
}
\description{
La fonction Tarif_EMERGE est un tarif de cubage à trois entrées proposé par l'ONF. 
Elle nécessite en entrée un data frame contenant par arbre sa hauteur totale (m),
circonférence (cm) et l'essence identifiée par le code IFN.
}
\examples{
df <- Tarif_EMERGE(df, df$Haut, df$C130)

}
\author{
Martin Marie-Laure
}
