% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vol.R
\docType{data}
\name{Vol}
\alias{Vol}
\title{Volumes.}
\format{
data frame

\describe{
  \item{Essence}{Nom de l'essence d'arbres}
  \item{Diam}{Diamètre (en cm)}
  \item{Vol}{Volume (en m3)}
}
}
\usage{
Vol
}
\description{
Exemple de volumes par essences
}
\examples{
data(Vol)

}
\keyword{datasets}
