% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolEchan.R
\docType{data}
\name{VolEchan}
\alias{VolEchan}
\title{Jeu de données de volume.}
\format{
data frame contenant 20 lignes et 2 variables
}
\usage{
VolEchan
}
\arguments{
\item{Diam}{= Diamètre en cm}

\item{Volume}{= Volume en m3}
}
\description{
Exemple de données diamètre volume pouvant être utilisé pour tester 
un tarif de cubage à 1 entrée
}
\examples{
data(VolEchan)


}
\keyword{datasets}
