% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolEmerge.R
\docType{data}
\name{VolEmerge}
\alias{VolEmerge}
\title{Coefficients par essence pour le tarif de cubage issu du projet EMERGE (volume total)}
\format{
data frame
}
\source{
C. Deleuze, F. Morneau, J.P. Renaud, Y. Vivien, M. Rivoire, et al.. 
Estimer le volume total d’un arbre, quelles que soient l’essence, la taille, la sylviculture, 
la station. Rendez-vous Techniques ONF, 2014, pp.22-32. hal-01143797
}
\usage{
VolEmerge
}
\arguments{
\item{CodeIFN}{= code de l'essence dans la base de données de l'IFN}

\item{Essence}{= essence pour lesquelles le modèle a été testé et validé statistiquement}

\item{A}{= sans unité, spécifique à l'essence}

\item{B}{= robustesse, commune pour tous les feuillus ou tous les résineux}

\item{C}{= défilement, sans unité, spécifique à l'essence}
}
\description{
Coefficients du tarif de cubage issu du projet EMERGE nécessaires pour calculer le volume total 
d'un arbre (volume total aérien, jusqu'à extrêmité des rameaux), par essence.
}
\examples{
data(VolEmerge)
Fonction associée Tarif_EMERGE() pour le calcul

}
\author{
Marie-Laure Martin
}
\keyword{datasets}
