% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolLot.R
\docType{data}
\name{VolLot}
\alias{VolLot}
\title{Lot d'arbres à cuber.}
\format{
data frame contenant 20 lignes et 2 variables
}
\usage{
VolLot
}
\arguments{
\item{Diam}{= Diamètre en cm}

\item{Nombre}{= Nombre de tiges par classe de diamètre}
}
\description{
Lot d'arbres à cuber.
}
\examples{
data(VolLot)

}
\keyword{datasets}
