% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodesEssIFN.R
\docType{data}
\name{CodesEssIFN}
\alias{CodesEssIFN}
\title{Codification des essences par l'IFN.}
\format{
Fichier contenant près de 142 lignes et 2 variables

\describe{
  \item{code}{Code de l'IFN}
  \item{libelle}{Nom vernaculaire}
}
}
\source{
\url{https://inventaire-forestier.ign.fr}
}
\usage{
CodesEssIFN
}
\description{
Table permettant de faire la correespondance entre le code IFN et le nom vernaculaire
}
\examples{
data(CodesEssIFN)

}
\keyword{datasets}
