% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decChaude.R
\docType{data}
\name{decChaude}
\alias{decChaude}
\title{Décroissances métriques des tarifs Chaudé.}
\format{
data frame contenant 19 lignes et 21 variables
}
\source{
Vade-mecum du forestier
}
\usage{
decChaude
}
\arguments{
\item{Classe}{= classe de diamètre en cm}

\item{Ch1}{à Ch20 = 20 numéros de tarifs Chaudé}
}
\description{
Coefficients des décroissances métriques variables exprimées en cm/m
pour les tarifs de cubage Chaudé à 2 entrées
}
\examples{
data(decChaude)

}
\keyword{datasets}
