% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNplacettes.R
\docType{data}
\name{IFNplacettes}
\alias{IFNplacettes}
\title{Fichier placettes de l'IFN.}
\format{
Data frame contenant près contenant 87535 lignes et 8 variables
}
\source{
\url{https://inventaire-forestier.ign.fr}
}
\usage{
IFNplacettes
}
\arguments{
\item{idp}{= identifiant du point d’inventaire}

\item{xl93, }{yl93 = coordonnées (latitude, longitude) en Lambert 93}

\item{ser}{= sylvoécorégion}

\item{csa}{= couverture du sol}

\item{dc}{= type de coupe}

\item{dist}{= distance de débardage en m}

\item{Annee}{= année de mesure}

\item{Type}{= type de climat}
}
\description{
localisation des placettes de l'IFN.
}
\examples{
data(IFNplacettes)

}
\keyword{datasets}
