% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSEEreg.R
\docType{data}
\name{INSEEreg}
\alias{INSEEreg}
\title{Régions françaises.}
\format{
Data frame contenant 18 lignes et 4 variables.
}
\source{
\url{https://www.insee.fr/fr/information/2666684#titre-bloc-26}
}
\usage{
INSEEreg
}
\arguments{
\item{REGION}{= Code région}

\item{CHEFLIEU}{= Chef-lieu de région}

\item{NCC}{= Nom en clair (majuscules)}

\item{NCCENR}{= Nom en clair (typographie riche)}
}
\description{
Liste des régions françaises existantes au 1er janvier 2017.
}
\examples{
data(INSEEreg)
 
}
\keyword{datasets}
