% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnIFN.R
\docType{data}
\name{rnIFN}
\alias{rnIFN}
\title{Régions naturelles de l'IFN.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 434 rows and 3 columns.
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?article729}
}
\usage{
rnIFN
}
\description{
Fichier géoréférencé correspondant à l'ancien découpage par l'IFN de la France métropolitaine
en 309 régions naturelles. Elles ont été remplacées par les sylvoécorégions.

\describe{
  \item{REGN}{identifiant de la région naturelle}
  \item{REGIONN}{intitulé de la région naturelle}
  \item{geometry}{géométrie}
}
}
\examples{
data(rnIFN)
 
}
\keyword{datasets}
