% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser.R
\docType{data}
\name{ser}
\alias{ser}
\title{Sylvoécorégions.}
\format{
Fichier géoréférencé contenant 86 sylvoécorégions.

\describe{
  \item{codeser}{identifiant du point d’inventaire}
  \item{NomSER, yl93}{coordonnées (latitude, longitude) en Lambert 93}
  \item{geometry}{géométrie}
}
}
\source{
\url{https://inventaire-forestier.ign.fr/spip.php?article729}
}
\usage{
ser
}
\description{
Carte des sylvoécorégions
}
\examples{
data(ser)

}
\keyword{datasets}
